file header info and test helper fix
