+++ packages/helpers/test/testing.test.js
@@ -1,0 +1,33 @@
+const { listFiles } = require('@asyncapi/generator-helpers');
+const fs = require('fs/promises');
+
+jest.mock('fs/promises');
+describe('listFiles', () => {
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('should return only file names from the directory', async () => {
+    const mockDirents = [
+      { name: 'file1.txt', isFile: () => true },
+      { name: 'file2.js', isFile: () => true },
+      { name: 'subdir', isFile: () => false },
+    ];
+
+    fs.readdir.mockResolvedValue(mockDirents);
+    const mockPath = '/mock/path';
+
+    const result = await listFiles(mockPath);
+    expect(fs.readdir).toHaveBeenCalledWith(mockPath, { withFileTypes: true });
+    expect(result).toEqual(['file1.txt', 'file2.js']);
+  });
+
+  it('should return an empty array if no files exist', async () => {
+    fs.readdir.mockResolvedValue([
+      { name: 'folder', isFile: () => false },
+    ]);
+    const mockPath = '/mock/path';
+    const result = await listFiles(mockPath);
+    expect(result).toEqual([]);
+  });
+});
