+++ packages/helpers/test/utils.test.js
@@ -1,7 +1,6 @@
 const path = require('path');
 const { Parser, fromFile } = require('@asyncapi/parser');
-const { getClientName, getInfo, getTitle, toSnakeCase, listFiles } = require('@asyncapi/generator-helpers');
-const fs = require('fs/promises');
+const { getClientName, getInfo, getTitle, toSnakeCase } = require('@asyncapi/generator-helpers');
 
 const parser = new Parser();
 const asyncapi_v3_path = path.resolve(__dirname, './__fixtures__/asyncapi-websocket-query.yml');
@@ -153,36 +152,5 @@ describe('toSnakeCase integration test with AsyncAPI', () => {
     const actualOperationId = toSnakeCase('');
     const expectedOperationId = '';
     expect(actualOperationId).toBe(expectedOperationId);
-  });
-});
-
-jest.mock('fs/promises');
-describe('listFiles', () => {
-  afterEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it('should return only file names from the directory', async () => {
-    const mockDirents = [
-      { name: 'file1.txt', isFile: () => true },
-      { name: 'file2.js', isFile: () => true },
-      { name: 'subdir', isFile: () => false },
-    ];
-
-    fs.readdir.mockResolvedValue(mockDirents);
-    const mockPath = '/mock/path';
-
-    const result = await listFiles(mockPath);
-    expect(fs.readdir).toHaveBeenCalledWith(mockPath, { withFileTypes: true });
-    expect(result).toEqual(['file1.txt', 'file2.js']);
-  });
-
-  it('should return an empty array if no files exist', async () => {
-    fs.readdir.mockResolvedValue([
-      { name: 'folder', isFile: () => false },
-    ]);
-    const mockPath = '/mock/path';
-    const result = await listFiles(mockPath);
-    expect(result).toEqual([]);
   });
 });
